feat: rename to AeIndexer
