+++ samples/TokenAeIndexer/Contracts/TokenContract.c.cs
