+++ samples/TokenAeIndexer/Entities/TransferRecord.cs
@@ -1,0 +1,12 @@
+using AeFinder.Sdk.Entities;
+using Nest;
+
+namespace TokenAeIndexer.Entities;
+
+public class TransferRecord: AeFinderEntity, IAeFinderEntity
+{
+    [Keyword] public string Symbol { get; set; }
+    [Keyword] public string FromAddress { get; set; }
+    [Keyword] public string ToAddress { get; set; }
+    public long Amount { get; set; }
+}
