+++ samples/TokenAeIndexer/GraphQL/AccountDto.cs
@@ -1,0 +1,10 @@
+using AeFinder.Sdk.Dtos;
+
+namespace TokenAeIndexer.GraphQL;
+
+public class AccountDto : AeFinderEntityDto
+{
+    public string Address { get; set; }
+    public string Symbol { get; set; }
+    public long Amount { get; set; }
+}
