+++ samples/TokenAeIndexer/GraphQL/AeIndexerSchema.cs
@@ -1,0 +1,10 @@
+using AeFinder.Sdk;
+
+namespace TokenAeIndexer.GraphQL;
+
+public class AeIndexerSchema : AppSchema<Query>
+{
+    public AeIndexerSchema(IServiceProvider serviceProvider) : base(serviceProvider)
+    {
+    }
+}
