+++ samples/TokenAeIndexer/GraphQL/TransferRecordDto.cs
@@ -1,0 +1,11 @@
+using AeFinder.Sdk.Dtos;
+
+namespace TokenAeIndexer.GraphQL;
+
+public class TransferRecordDto : AeFinderEntityDto
+{
+    public string Symbol { get; set; }
+    public string FromAddress { get; set; }
+    public string ToAddress { get; set; }
+    public long Amount { get; set; }
+}
