--- samples/TokenApp/Entities/Account.cs
@@ -1,11 +1,0 @@
-using AeFinder.Sdk.Entities;
-using Nest;
-
-namespace TokenApp.Entities;
-
-public class Account: AeFinderEntity, IAeFinderEntity
-{
-    [Keyword] public string Address { get; set; }
-    [Keyword] public string Symbol { get; set; }
-    public long Amount { get; set; }
-}
