--- samples/TokenApp/GraphQL/AccountDto.cs
@@ -1,10 +1,0 @@
-using AeFinder.Sdk.Dtos;
-
-namespace TokenApp.GraphQL;
-
-public class AccountDto : AeFinderEntityDto
-{
-    public string Address { get; set; }
-    public string Symbol { get; set; }
-    public long Amount { get; set; }
-}
