--- samples/TokenApp/GraphQL/AppSchema.cs
@@ -1,10 +1,0 @@
-using AeFinder.Sdk;
-
-namespace TokenApp.GraphQL;
-
-public class AppSchema : AppSchema<Query>
-{
-    public AppSchema(IServiceProvider serviceProvider) : base(serviceProvider)
-    {
-    }
-}
