--- samples/TokenApp/GraphQL/GetTransferRecordInput.cs
@@ -1,8 +1,0 @@
-namespace TokenApp.GraphQL;
-
-public class GetTransferRecordInput
-{
-    public string ChainId { get; set; }
-    public string Address { get; set; }
-    public string Symbol { get; set; }
-}
