--- samples/TokenApp/GraphQL/TransferRecordDto.cs
@@ -1,11 +1,0 @@
-using AeFinder.Sdk.Dtos;
-
-namespace TokenApp.GraphQL;
-
-public class TransferRecordDto : AeFinderEntityDto
-{
-    public string Symbol { get; set; }
-    public string FromAddress { get; set; }
-    public string ToAddress { get; set; }
-    public long Amount { get; set; }
-}
