+++ src/AeFinder.Application/DevelopmentTemplate/DevTemplateOptions.cs
@@ -5,7 +5,7 @@ namespace AeFinder.DevelopmentTemplate;
 public class DevTemplateOptions
 {
     public string TemplatePath { get; set; } = "DevTemplate";
-    public string ProjectPlaceholder { get; set; } = "AppTemplate";
+    public string ProjectPlaceholder { get; set; } = "AeIndexerTemplate";
     public List<string> ReplaceExtensions { get; set; } = new() { ".cs", ".csproj", ".sln" };
     public List<string> ProjectExtensions { get; set; } = new() { ".csproj", ".sln" };
