+++ templates/AeFinderAppTemplate/AeIndexerTemplate.sln
@@ -1,0 +1,30 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{F07E638D-DEC9-4D0B-A815-791C0AED81C0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AeIndexerTemplate", "src\AeIndexerTemplate\AeIndexerTemplate.csproj", "{C00C5349-FF95-41F5-A56B-299986F157EB}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{CF8ECC12-55BE-457D-A3EE-D4283041B686}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AeIndexerTemplate.Tests", "test\AeIndexerTemplate.Tests\AeIndexerTemplate.Tests.csproj", "{1DA69092-9C61-4660-8A61-3BA30ED892E0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{C00C5349-FF95-41F5-A56B-299986F157EB} = {F07E638D-DEC9-4D0B-A815-791C0AED81C0}
+		{1DA69092-9C61-4660-8A61-3BA30ED892E0} = {CF8ECC12-55BE-457D-A3EE-D4283041B686}
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C00C5349-FF95-41F5-A56B-299986F157EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C00C5349-FF95-41F5-A56B-299986F157EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C00C5349-FF95-41F5-A56B-299986F157EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C00C5349-FF95-41F5-A56B-299986F157EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1DA69092-9C61-4660-8A61-3BA30ED892E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1DA69092-9C61-4660-8A61-3BA30ED892E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1DA69092-9C61-4660-8A61-3BA30ED892E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1DA69092-9C61-4660-8A61-3BA30ED892E0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
