+++ templates/AeFinderAppTemplate/src/AeIndexerTemplate/AeIndexerTemplateProfile.cs
@@ -1,0 +1,13 @@
+using AeIndexerTemplate.Entities;
+using AeIndexerTemplate.GraphQL;
+using AutoMapper;
+
+namespace AeIndexerTemplate;
+
+public class AeIndexerTemplateProfile : Profile
+{
+    public AeIndexerTemplateProfile()
+    {
+        CreateMap<MyEntity, MyEntityDto>();
+    }
+}
