+++ templates/AeFinderAppTemplate/src/AeIndexerTemplate/Entities/MyEntity.cs
@@ -1,0 +1,11 @@
+using AeFinder.Sdk.Entities;
+using Nest;
+
+namespace AeIndexerTemplate.Entities;
+
+public class MyEntity: AeFinderEntity, IAeFinderEntity
+{
+    [Keyword] public string Address { get; set; }
+    [Keyword] public string Symbol { get; set; }
+    public long Amount { get; set; }
+}
