+++ templates/AeFinderAppTemplate/src/AeIndexerTemplate/GraphQL/GetMyEntityInput.cs
@@ -1,0 +1,8 @@
+namespace AeIndexerTemplate.GraphQL;
+
+public class GetMyEntityInput
+{
+    public string ChainId { get; set; }
+    public string Address { get; set; }
+    public string Symbol { get; set; }
+}
