+++ templates/AeFinderAppTemplate/src/AeIndexerTemplate/GraphQL/MyEntityDto.cs
@@ -1,0 +1,10 @@
+using AeFinder.Sdk.Dtos;
+
+namespace AeIndexerTemplate.GraphQL;
+
+public class MyEntityDto : AeFinderEntityDto
+{
+    public string Address { get; set; }
+    public string Symbol { get; set; }
+    public long Amount { get; set; }
+}
