--- templates/AeFinderAppTemplate/src/AppTemplate/AppTemplateProfile.cs
@@ -1,13 +1,0 @@
-using AppTemplate.Entities;
-using AppTemplate.GraphQL;
-using AutoMapper;
-
-namespace AppTemplate;
-
-public class AppTemplateProfile : Profile
-{
-    public AppTemplateProfile()
-    {
-        CreateMap<MyEntity, MyEntityDto>();
-    }
-}
