--- templates/AeFinderAppTemplate/src/AppTemplate/GraphQL/GetMyEntityInput.cs
@@ -1,8 +1,0 @@
-namespace AppTemplate.GraphQL;
-
-public class GetMyEntityInput
-{
-    public string ChainId { get; set; }
-    public string Address { get; set; }
-    public string Symbol { get; set; }
-}
