+++ templates/AeFinderAppTemplate/test/AeIndexerTemplate.Tests/AeIndexerTemplateTestBase.cs
@@ -1,0 +1,8 @@
+using AeFinder.App.TestBase;
+
+namespace AeIndexerTemplate;
+
+public abstract class AeIndexerTemplateTestBase: AeFinderAppTestBase<AeIndexerTemplateTestModule>
+{
+
+}
