+++ templates/AeFinderAppTemplate/test/AeIndexerTemplate.Tests/AssemblyInfo.cs
