--- templates/AeFinderAppTemplate/test/AppTemplate.Tests/AppTemplate.Tests.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <RootNamespace>AppTemplate</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\AppTemplate\AppTemplate.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="AeFinder.App.TestBase" Version="1.2.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
-        <PackageReference Include="xunit" Version="2.6.1" />
-        <PackageReference Include="xunit.extensibility.execution" Version="2.6.1" />
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
-        <PackageReference Include="Shouldly" Version="4.2.1" />
-        <PackageReference Include="coverlet.collector" Version="6.0.2">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-    </ItemGroup>
-</Project>
