--- templates/AeFinderAppTemplate/test/AppTemplate.Tests/AppTemplateTestBase.cs
@@ -1,8 +1,0 @@
-using AeFinder.App.TestBase;
-
-namespace AppTemplate;
-
-public abstract class AppTemplateTestBase: AeFinderAppTestBase<AppTemplateTestModule>
-{
-
-}
