--- templates/AeFinderAppTemplate/test/AppTemplate.Tests/AppTemplateTestModule.cs
@@ -1,19 +1,0 @@
-using AeFinder.App.TestBase;
-using Microsoft.Extensions.DependencyInjection;
-using Volo.Abp.Modularity;
-
-namespace AppTemplate;
-
-[DependsOn(
-    typeof(AeFinderAppTestBaseModule),
-    typeof(AppTemplateModule))]
-public class AppTemplateTestModule : AbpModule
-{
-    public override void ConfigureServices(ServiceConfigurationContext context)
-    {
-        Configure<AeFinderAppEntityOptions>(options => { options.AddTypes<AppTemplateModule>(); });
-        
-        // Add your Processors.
-        // context.Services.AddSingleton<MyLogEventProcessor>();
-    }
-}
